# dist/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CRUD API",
        "description": "Essa API simula um CRUD dos usuarios",
        "contact": {
            "email": "lucas.bernardes.santos.mail@gmail.com"
        },
        "version": 1.0
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "API de teste"
        },
        {
            "url": "http://www.usuarios.com",
            "description": "API de produção"
        }
    ],
    "paths": {
        "/usuario/cadastrar": {
            "post": {
                "summary": "Cadastrar usuário",
                "description": "Rota responsável por cadastrar novos usuários",
                "tags": [
                    "Usuario"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Usuario"
                            },
                            "examples": {
                                "usuario": {
                                    "value": {
                                        "nome": "Liz Adriana dos Santos",
                                        "cpf": "13346563715",
                                        "nascimento": "1996/05/21",
                                        "sexo": "F",
                                        "saldo": 3549.78
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/Usuario"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Bad Request - Mensagem de Erro"
                    }
                }
            }
        },
        "/usuario/buscar/{cpf}": {
            "get": {
                "summary": "Buscar usuário",
                "description": "Buscar um usuário apartir do cpf enviado",
                "tags": [
                    "Usuario"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "cpf",
                        "schema": {
                            "type": "string",
                            "example": "17022834709"
                        },
                        "required": true,
                        "description": "CPF do usuário, usado para identifica-lo no banco"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/Usuario"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/usuario/atualizar": {
            "put": {
                "summary": "Atualizar usuário",
                "description": "Rota responsável por alterar os dados de um usuário",
                "tags": [
                    "Usuario"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Usuario"
                            },
                            "examples": {
                                "usuario": {
                                    "value": {
                                        "nome": "Liz Adriana dos Santos",
                                        "cpf": "13346563715",
                                        "nascimento": "1996/05/21",
                                        "sexo": "F",
                                        "saldo": 3549.78
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "number",
                                    "description": "Quantidade de linhas alteradas",
                                    "example": 1
                                }
                            }
                        }
                    }
                }
            }
        },
        "/usuario/deletar/{cpf}": {
            "delete": {
                "summary": "Deletar usuário",
                "description": "Deletar um usuário apartir do cpf enviado",
                "tags": [
                    "Usuario"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "cpf",
                        "schema": {
                            "type": "string",
                            "example": "17022834709"
                        },
                        "required": true,
                        "description": "CPF do usuário, usado para identifica-lo no banco"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "number",
                                    "description": "Quantidade de linhas alteradas",
                                    "example": 0
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Usuario": {
                "type": "object",
                "properties": {
                    "nome": {
                        "type": "string"
                    },
                    "cpf": {
                        "type": "string"
                    },
                    "nascimento": {
                        "type": "string",
                        "format": "date"
                    },
                    "sexo": {
                        "type": "string",
                        "enum": [
                            "F",
                            "M"
                        ]
                    },
                    "saldo": {
                        "type": "number",
                        "format": "float",
                        "example": 1.25
                    }
                }
            }
        }
    }
}